Fixed Duplicate File Issue
When processing sub directories there was some issues with duplicated variable names. So i replaced them with generic FILE_0 ... names in the generated D code. 

I also fixed the depreciated call to outputBin2D2FS when creating a temporary folder.

Added a cleanup function that removes the directory. Not really sure if I should... but it was useful to me.

changed some of the double quotes to ticks to take advantage of D's wysiwyg feature.

Also I am quite new to GIT in general if I have done the wrong thing just let me know :)